(LEX): some improvements and bugfixes:  -  do not evaluate token type each time it is requested  -  extract some boilerplate to `LexerBaseEx`  -  fix awful bug in `RustEscapesLexer`
