+++ src/main/kotlin/org/rust/lang/core/lexer/RustEscapesLexer.kt
@@ -4,6 +4,7 @@ import com.intellij.openapi.util.text.StringUtil
 import com.intellij.psi.StringEscapesTokenTypes.*
 import com.intellij.psi.tree.IElementType
 import org.rust.lang.core.psi.RustTokenElementTypes.*
+import org.rust.lang.utils.isRustWhitespaceChar
 
 private const val BYTE_ESCAPE_LENGTH = "\\x00".length
 private const val UNICODE_ESCAPE_MIN_LENGTH = "\\u{0}".length
@@ -14,22 +15,8 @@ private const val UNICODE_ESCAPE_MAX_LENGTH = "\\u{000000}".length
  */
 class RustEscapesLexer private constructor(val defaultToken: IElementType,
                                            val unicode: Boolean = false,
-                                           val eol: Boolean = false) : LexerBaseKt() {
-    override fun start(buffer: CharSequence, startOffset: Int, endOffset: Int, initialState: Int) {
-        bufferSequence = buffer
-        bufferEnd = endOffset
-        state = initialState
-
-        tokenStart = startOffset
-        tokenEnd = locateToken(tokenStart)
-    }
-
-    override fun advance() {
-        tokenStart = tokenEnd
-        tokenEnd = locateToken(tokenStart)
-    }
-
-    override fun getTokenType(): IElementType? {
+                                           val eol: Boolean = false) : LexerBaseEx() {
+    protected override fun determineTokenType(): IElementType? {
         // We're at the end of the string token => finish lexing
         if (tokenStart >= tokenEnd) {
             return null
@@ -48,18 +35,18 @@ class RustEscapesLexer private constructor(val defaultToken: IElementType,
         return when (bufferSequence[tokenStart + 1]) {
             'u' ->
                 when {
-                    !unicode -> INVALID_CHARACTER_ESCAPE_TOKEN
+                    !unicode                                   -> INVALID_CHARACTER_ESCAPE_TOKEN
                     isValidUnicodeEscape(tokenStart, tokenEnd) -> VALID_STRING_ESCAPE_TOKEN
-                    else -> INVALID_UNICODE_ESCAPE_TOKEN
+                    else                                       -> INVALID_UNICODE_ESCAPE_TOKEN
                 }
-            'x' -> esc(isValidByteEscape(tokenStart, tokenEnd))
-            '\r', '\n' -> esc(eol)
+            'x'                                 -> esc(isValidByteEscape(tokenStart, tokenEnd))
+            '\r', '\n'                          -> esc(eol)
             'n', 'r', 't', '0', '\\', '\'', '"' -> VALID_STRING_ESCAPE_TOKEN
-            else -> INVALID_CHARACTER_ESCAPE_TOKEN
+            else                                -> INVALID_CHARACTER_ESCAPE_TOKEN
         }
     }
 
-    private fun locateToken(start: Int): Int {
+    protected override fun locateToken(start: Int): Int {
         if (start >= bufferEnd) {
             return start
         }
@@ -85,11 +72,13 @@ class RustEscapesLexer private constructor(val defaultToken: IElementType,
                         val idx = bufferSequence.indexOf('}', i + 1)
                         return if (idx != -1) Math.min(idx + 1, bufferEnd) else bufferEnd
                     }
-                'r' ->
-                    // Check if we have \r\n and consume additional \n
-                    if (bufferEnd - (i + 1) >= 2 && bufferSequence.startsWith("\\n", i + 1)) {
-                        return i + 2 + 1
+                '\r', '\n' -> {
+                    var j = i
+                    while (j < bufferEnd && bufferSequence[j].isRustWhitespaceChar()) {
+                        j++
                     }
+                    return j
+                }
             }
             return i + 1
         } else {
@@ -142,11 +131,11 @@ class RustEscapesLexer private constructor(val defaultToken: IElementType,
          * @throws IllegalArgumentException when given token type is unsupported
          */
         fun of(tokenType: IElementType): RustEscapesLexer = when (tokenType) {
-            BYTE_LITERAL -> forByteLiterals()
-            CHAR_LITERAL -> forCharLiterals()
+            BYTE_LITERAL        -> forByteLiterals()
+            CHAR_LITERAL        -> forCharLiterals()
             BYTE_STRING_LITERAL -> forByteStringLiterals()
-            STRING_LITERAL -> forStringLiterals()
-            else -> throw IllegalArgumentException("unsupported literal type")
+            STRING_LITERAL      -> forStringLiterals()
+            else                -> throw IllegalArgumentException("unsupported literal type")
         }
     }
 }
