+++ src/main/kotlin/org/rust/lang/utils/CharEx.kt
@@ -1,0 +1,6 @@
+package org.rust.lang.utils
+
+/**
+ * Determines if the char is a Rust whitespace character.
+ */
+fun Char.isRustWhitespaceChar(): Boolean = equals(' ') || equals('\r') || equals('\n') || equals('\t')
