+++ src/test/resources/org/rust/lang/core/lexer/fixtures/highlighting/eol.txt
@@ -16,8 +16,8 @@ WHITE_SPACE (' ')
 = ('=')
 WHITE_SPACE (' ')
 <STRING> ('"foo')
-VALID_STRING_ESCAPE_TOKEN ('\\n')
-<STRING> ('         bar"')
+VALID_STRING_ESCAPE_TOKEN ('\\n         ')
+<STRING> ('bar"')
 ; (';')
 WHITE_SPACE ('\n\n')
 <IDENTIFIER> ('assert_eq')
